Add torrent links to RemoteAddon model
This modifications extend the RemoteAddon model to feature torrent links. The 'link' attribute was changed to 'page' and a new attribute 'torrent' was added to define specific download links for each RemoteAddon instance. Also, a 'download' function was created to provide immediate redirection to these links.
Add actions for 'view' and 'download' were also incorporated to facilitate user access.