Refactor display of remote addon descriptions
The previous implementation of remote addon descriptions was removed from the table columns and added as an action. This change improves the user interface by preventing long descriptions from crowding the main table and instead displays them upon user interaction with the 'info' button. This makes for a much cleaner overview, with detailed information still accessible.