Update matchLocalAddons method and remove loilo/fuse package
The matchLocalAddons method in LocalAddon model was modified to match local addons with the remote ones not only based on the similar title but also considering the version and author. The fuzzy search library 'loilo/fuse' was not needed anymore, hence it was removed from composer.json file.

This approach was taken to avoid incorrect matches that can occur when titles are similar but authors and versions are different, ensuring the matching process is more accurate. The code related to the 'loilo/fuse' package was also removed from composer.lock and the LocalAddonResource file was updated to reflect the changes in matching logic.