Fix version comparison in LocalAddon model and resource
The version comparison in LocalAddon model and resource was modified. Previously, the comparison included the tailing ".0" in version numbers, which resulted in false positives when comparing versions where the only difference was the existence of this tailing ".0". After the change, the ".0" is removed from the version numbers before comparison, ensuring more accurate comparison results.