Refactor RemoteAddon model and update time column in Resource
The RemoteAddon model has been refactored. Changes include the addition of Carbon for manipulating date and time. The new and old addons now have their created_at and updated_at attributes set with the current time upon creation or update. Also, in the RemoteAddonResource, the date column is changed from 'updated_at' to 'published_at'. This provides a more accurate representation of the addons' status, improving usability and data accuracy.