Add NativePHP
This commit adds the NativeAppServiceProvider and a nativephp configuration. The NativeAppServiceProvider sets up the native application bindings and configurations, making it easier for developers to customize and control native applications. The nativephp configuration file provides a standard way for defining and accessing essential app settings, such as the app version, id, author, etc. This change also includes updates in `composer.json` and `.gitignore` files.