Update startup sequence in release workflow
Modified the order of commands within the startup batch file in the Github release workflow.  The commands to run PHP artisan serve and to open the browser to localhost were previously intertwined. They are now separated such that the server is started before attempting to open the browser. This ensures the server is running before a browser request is made, avoiding potential race conditions and subsequent user confusion.