Remove Application Caching in GitHub Workflow
In this commit, the application caching section in GitHub workflow has been commented out. This action was necessary to avoid potential issues in app deployment due to stale cache. This change doesn't impact the function of the other parts of the workflow. Also, the path for php execution in Windows has been updated to use forward slash, which is more compatible and could avoid potential issues.