Implement FilamentUser interface in User model
This commit implements the FilamentUser interface in the User model and adds the canAccessPanel function. This change was made to allow every user to access the control panel. In .github/workflows/release.yml, the order of the 'echo' commands in .bat script has been swapped, ensuring that the server starts after the webpage has finished loading to prevent any loading issues.