Refactor warning and description extraction, enhance error handling
Refactored the extraction logic of 'warning' and 'description' in RemoteAddon model to use Laravel's Str class, improving readability and maintainability. Also added an error handling step on LocalAddon to save remote add-ons when none are found. Removed static pages and widgets declaration in AdminPanelProvider; these are now managed by the discovery method.

Minor trims were made in the RemoteAddonResource and enabled the 'intl' extension in php.ini for i18n feature. These changes are part of the effort to streamline the project codebase and enhance its overall robustness.