Refactor 'is_updated' calculation logic for LocalAddon model
Replaced the direct 'is_updated' boolean field in the LocalAddon model with a computed attribute that calculates the value based on version comparison between local and remote addons. This change was necessary because the 'is_updated' field was not reactive to version changes in related models. Now the 'is_updated' attribute will always reflect the most current state.

Also, in this commit, the 'info' Action was improved to show detail information about an addon when not null, and extraAttributes logic was reactivated to apply different classes based on the remoteAddon description existence. These changes will make UI more informative and intuitive to use.