Refactor recommendation options in LocalAddon and RemoteAddon Resources
The options for the 'is_recommended' filter have been refactored. Instead of explicitly stating the options in an array, these have been replaced with the toArray method of the IsRecommended enum class. This change makes the code cleaner and easier to maintain. An indicateUsing function is also added in LocalAddonResource to display the recommendation status. Additionally, URL generation for 'view_matching_addon' has been replaced with page generation.